Refactor exception handling in player commands and update level handler initialization

+++ src/main/java/org/me/newsky/island/PlayerHandler.java
@@ -45,6 +45,11 @@ public class PlayerHandler {
                 throw new IslandPlayerDoesNotExistException();
             }
 
+            UUID ownerUuid = cacheHandler.getIslandOwner(islandUuid);
+            if (ownerUuid.equals(playerUuid)) {
+                throw new CannotRemoveOwnerException();
+            }
+
             cacheHandler.deleteIslandPlayer(islandUuid, playerUuid);
         }, plugin.getBukkitAsyncExecutor());
     }
